Start adding testing that uses the CloudNative impl
