+++ akka-remote/src/main/resources/reference.conf
@@ -1107,7 +1107,7 @@ akka {
         # Must implement akka.remote.artery.tcp.SSLEngineProvider and have a public
         # constructor with an ActorSystem parameter.
         # The default ConfigSSLEngineProvider is configured by properties in section
-        # akka.remote.artery.ssl.config-ssl-engine
+        # akka.remote.artery.tcp.ssl.config-ssl-engine
         ssl-engine-provider = akka.remote.artery.tcp.ssl.ConfigSSLEngineProvider
 
         # Config of akka.remote.artery.tcp.ssl.ConfigSSLEngineProvider
@@ -1178,7 +1178,8 @@ akka {
           # How long should an SSLContext instance be cached. When rotating keys and certificates,
           # there must a time overlap between the old certificate and the new one. This value should
           # be lower than the certificates overlap.
-          ssl-context-cache-ttl = 1h
+          # Defaults to 500 days (for compatibilty with the non expiring, legacy implementation)
+          ssl-context-cache-ttl = 500d
 
         }
 
@@ -1214,7 +1215,7 @@ akka {
           ca-cert-file = ${akka.remote.artery.ssl.cloud-native-ssl-engine.secret-mount-point}/ca.crt
 
           # Protocol to use for SSL encryption, choose from:
-          protocol = "TLSv1.2"
+          protocol = ${akka.remote.artery.ssl.config-ssl-engine.protocol}
 
           # Example: ["TLS_DHE_RSA_WITH_AES_128_GCM_SHA256",
           #   "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
@@ -1224,7 +1225,7 @@ akka {
           # the JCE Unlimited Strength Jurisdiction Policy Files to use AES 256.
           # More info here:
           # https://www.oracle.com/java/technologies/javase-jce-all-downloads.html
-          enabled-algorithms = ["TLS_DHE_RSA_WITH_AES_256_GCM_SHA384"]
+          enabled-algorithms = ${akka.remote.artery.ssl.config-ssl-engine.enabled-algorithms}
 
           # There are two options, and the default SecureRandom is recommended:
           # "" or "SecureRandom" => (default)
@@ -1232,7 +1233,7 @@ akka {
           #
           # Setting a value here may require you to supply the appropriate cipher
           # suite (see enabled-algorithms section above)
-          random-number-generator = ""
+          random-number-generator = ${akka.remote.artery.ssl.config-ssl-engine.random-number-generator}
 
           # Require mutual authentication between TLS peers
           #
