+++ akka-remote/src/test/scala/akka/remote/artery/ArterySpecSupport.scala
@@ -32,6 +32,9 @@ object ArterySpecSupport {
   // set the test key-store and trust-store properties
   // TLS only used if transport=tls-tcp, which can be set from specific tests or
   // System properties (e.g. jenkins job)
+  // TODO: randomly return a Config using ConfigSSLEngineProvider or
+  //  CloudNativeSSLEngineProvider and, eventually, run tests twice
+  //  (once for each provider).
   lazy val tlsConfig: Config = {
     val trustStore = getClass.getClassLoader.getResource("truststore").getPath
     val keyStore = getClass.getClassLoader.getResource("keystore").getPath
